pyproject.toml
[project]
name = "server-backup"
version = "0.1.0"
description = "Server backup tool with Google Drive integration"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]
readme = "README.md"
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backup"]

[tool.uv]
dev-dependencies = []
